rename/move class
